+++ packages/core/src/history/index.ts
@@ -1,0 +1,2 @@
+export * from './manager';
+export * from './manager-default';
